+++ Authentication/controllers/googleAuth.js
@@ -1,4 +1,4 @@
-const {pool} = require('../dbConfig')
+const {pool} = require('../models/configDB')
 const jwt = require('jsonwebtoken');
 const JWT_SECRET = process.env.JWT_SECRET
 const { OAuth2Client } = require('google-auth-library');
