+++ Authentication/middleware/fetchuser.js
@@ -17,4 +17,4 @@ const fetchuser = (req, res, next) => {
     
 }
 
-module.exports = fetchuser
+module.exports = {fetchuser}
